# server/out/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/system/post/list": {
      "get": {
        "operationId": "PostController_findAll",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "当前分页(前端别名 page)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量(前端别名 limit)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "params",
            "required": false,
            "in": "query",
            "description": "时间范围",
            "schema": {
              "$ref": "#/components/schemas/DateParamsDTO"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "岗位列表",
        "tags": [
          "岗位管理"
        ]
      }
    },
    "/api/system/post/enabled": {
      "get": {
        "operationId": "PostController_enabled",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "启用的岗位",
        "tags": [
          "岗位管理"
        ]
      }
    },
    "/api/system/post/detail/{id}": {
      "get": {
        "operationId": "PostController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "岗位详情",
        "tags": [
          "岗位管理"
        ]
      }
    },
    "/api/system/post/create": {
      "post": {
        "operationId": "PostController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "创建岗位",
        "tags": [
          "岗位管理"
        ]
      }
    },
    "/api/system/post/update/{id}": {
      "put": {
        "operationId": "PostController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "更新岗位",
        "tags": [
          "岗位管理"
        ]
      }
    },
    "/api/system/post/delete/{id}": {
      "delete": {
        "operationId": "PostController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "删除岗位",
        "tags": [
          "岗位管理"
        ]
      }
    },
    "/api/system/post/status/{id}": {
      "put": {
        "operationId": "PostController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "更新岗位状态",
        "tags": [
          "岗位管理"
        ]
      }
    },
    "/api/system/post/access-post": {
      "get": {
        "operationId": "PostController_accessPost",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "可访问岗位",
        "tags": [
          "岗位管理"
        ]
      }
    },
...
    "/api/core/server/redis": {
      "get": {
        "operationId": "CacheController_getInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Redis监控信息",
        "tags": [
          "缓存管理"
        ]
      }
    },
    "/api/core/server/redis/browser/level1": {
      "get": {
        "operationId": "CacheController_getLevel1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Redis浏览器-一级目录",
        "tags": [
          "缓存管理"
        ]
      }
    },
    "/api/core/server/redis/browser/level2": {
      "get": {
        "operationId": "CacheController_getLevel2",
        "parameters": [
          {
            "name": "prefix",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Redis浏览器-二级目录",
        "tags": [
          "缓存管理"
        ]
      }
    },
    "/api/core/server/redis/browser/level3": {
      "get": {
        "operationId": "CacheController_getLevel3",
        "parameters": [
          {
            "name": "prefix",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Redis浏览器-三级目录",
        "tags": [
          "缓存管理"
        ]
      }
    },
    "/api/core/server/redis/browser/key-info": {
      "get": {
        "operationId": "CacheController_getKeyInfo",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Redis浏览器-Key信息",
        "tags": [
          "缓存管理"
        ]
      }
    },
    "/api/core/server/redis/browser/delete": {
      "delete": {
        "operationId": "CacheController_deleteKey",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Redis浏览器-删除Key",
        "tags": [
          "缓存管理"
        ]
      }
    },
...
    "/api/ai/admin/providers/list": {
      "get": {
        "operationId": "AiAdminController_listProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "供应商列表",
        "tags": [
          "AI 管理"
        ]
      }
    },
    "/api/ai/admin/providers/create": {
      "post": {
        "operationId": "AiAdminController_createProvider",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "创建供应商",
        "tags": [
          "AI 管理"
        ]
      }
    },
    "/api/ai/admin/providers/update/{id}": {
      "put": {
        "operationId": "AiAdminController_updateProvider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "更新供应商",
        "tags": [
          "AI 管理"
        ]
      }
    },
    "/api/ai/admin/providers/delete/{id}": {
      "delete": {
        "operationId": "AiAdminController_deleteProvider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "删除供应商",
        "tags": [
          "AI 管理"
        ]
      }
    },
    "/api/ai/admin/providers/options": {
      "get": {
        "operationId": "AiAdminController_providerOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "供应商下拉",
        "tags": [
          "AI 管理"
        ]
      }
    },
    "/api/ai/admin/models/list": {
      "get": {
        "operationId": "AiAdminController_listModels",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "模型列表",
        "tags": [
          "AI 管理"
        ]
      }
    },
    "/api/ai/admin/models/create": {
      "post": {
        "operationId": "AiAdminController_createModel",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "创建模型",
        "tags": [
          "AI 管理"
        ]
      }
    },
    "/api/ai/admin/models/update/{id}": {
      "put": {
        "operationId": "AiAdminController_updateModel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "更新模型",
        "tags": [
          "AI 管理"
        ]
      }
    },
    "/api/ai/admin/models/delete/{id}": {
      "delete": {
        "operationId": "AiAdminController_deleteModel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "删除模型",
        "tags": [
          "AI 管理"
        ]
      }
    }
  },
  "info": {
    "title": "FssAdmin",
    "description": "FssAdmin API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
      "DateParamsDTO": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "sort": {
            "type": "number"
          },
          "status": {
            "type": "number"
          },
          "remark": {
            "type": "string"
          }
        }
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "sort": {
            "type": "number"
          },
          "status": {
            "type": "number"
          },
          "remark": {
            "type": "string"
          },
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}